Implement q_ascend and q_descend
In q_ascend, the function iterates over the linked list from tail to
head and compare the value of current node and remove it if the value is
larger than the minimum value. Otherwise, keep if and update the minimum
value. q_descend does the opposite.
Because the two functions do almost the same thing, so I extract the
common part to a function q_ascend_descend() to improve readability.

Change-Id: I9dfb87cf7d097ab7beb915712e58e830bae6a17f